소스 정리
source clean

+++ app/src/main/java/de/mud/terminal/Precomposer.java
@@ -1032,12 +1032,12 @@ public class Precomposer {
 		int max = precompositions.length - 1;
 		int mid;
 
-		long sought = base << UNICODE_SHIFT | comb;
+		long sought = (long) base << UNICODE_SHIFT | comb;
 		long that;
 
 		while (max >= min) {
 			mid = (min + max) / 2;
-			that = precompositions[mid][1] << UNICODE_SHIFT | precompositions[mid][2];
+			that = (long) precompositions[mid][1] << UNICODE_SHIFT | precompositions[mid][2];
 			if (that < sought)
 				min = mid + 1;
 			else if (that > sought)
